Use .h5 file to process spectra, then extract pixels and write a .csv
I downloaded the .h5 files locally, but the neonUtilities package can be used to pull the files directly for the 9 stands in this study.  Spectra processing is shown here as an example for one stand.  Stands were manually adjusted to make 9 csv files, which are gathered in the subsequent script.