Alex uses plot centers now for downloading tiles.
Alex worked throught the code within the loop, but cannot get the loop to run.

The loop is looping through the .h5 files and associated dsm with nami?   I don't know what order the loop is in, but I think there needs to be an improved way of doing stand level processing which might be done by looking through the stand centers, and using the stand center to or extent to inform which .h5 file to "activate" for the processing.

Shade Mask looks great!

Please let me know if I have adequately account for working directories and feel free to erase old notes that I took while I was going through the code. I often put a date, for my own sanity but I'm all for keeping the code clean without extraneous comments.