Made new graphs from the PLSDA and ran ANOVA
Still need to make ANOVA table from the model.

Results didn't strongly change.  We are going with the idea that we are using individual tree tops for any analysis (PLSDA and mixed effect model N*P anova) and are using plot level average spectra for graphs because we incorporate field measurements that are made at the plot level.